Added an example with view
